[INJICERT-187] seperated VCIssuance to new module for certify (#28)
* [INJICERT-187] seperated VCIssuance to new module for certify (#26)

* [INJICERT-187] seperated VCIssuance to new module for certify

* [INJICERT-187] removed unused fields from audit dto

* [INJICERT-187] removed audit plugin implementation and changed the certify dependency version

* [DSD-5387]

* [DSD-5387]

---------
